requirements.txt
bcrypt==3.2.0
bidict==0.21.3
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cloudinary==1.26.0
colorama==0.4.4
cryptography==3.4.8
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
h11==0.12.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsondiff==1.3.0
MarkupSafe==2.0.1
mongoengine==0.23.1
ndg-httpsclient==0.5.1
pipi==1.0.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
pymongo==3.12.0
pymongo-model==2.0.2
pyOpenSSL==20.0.1
python-dotenv==0.19.0
python-engineio==4.2.1
python-socketio==5.4.0
requests==2.26.0
simple-websocket==0.4.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.1
wsproto==1.0.0
